refactor: extract header logic into custom hook for improved readability and maintainability
